+++ src/stores/index.ts
@@ -1,2 +1,2 @@
-export { buttonHandlerStore, BTN_TYPE } from '@src/stores/buttonHandler.store.ts';
-export { commonPrivateKeyStore } from '@src/stores/commonPrivateKey.store.ts';
+export { buttonHandlerStore, BTN_TYPE } from '@/stores/buttonHandler.store.ts';
+export { commonPrivateKeyStore } from '@/stores/commonPrivateKey.store.ts';
