+++ src/utils/http/axios/index.ts
@@ -1,4 +1,4 @@
-import { AxiosHttpClient } from '@src/utils/http/axios/axios.ts';
+import { AxiosHttpClient } from './axios';
 
 export function createAxios() {
   console.log(import.meta.env.VITE_API_SERVER_URL);
