+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { Char } from './char';
 export { Time } from './time';
 export { Crypto } from './crypto';
+export { service } from './http';
