+++ tsconfig.json
@@ -20,7 +20,7 @@
 
     "baseUrl": ".",
     "paths": {
-      "@src/*": ["src/*"],
+      "@/*": ["src/*"],
       "@hooks/*": ["src/hooks/*"],
       "@services/*": ["src/services/*"],
       "@components/*": ["src/components/*"],
